Remove shutter from tray, remove floating volume mixer
Most of the time I accidentally shut Shutter down, so the
tray icon wasn't consistent. It's just easier to run it when
needed.

Also floating mixer was a bad idea, since I like to have
my mixer open all the time at one workspace, and when it
was floating, it took the whole workspace.